Add support for git-managed projects
Source packages for some projects are now managed in git. lubed uses a
list of such projects to either ask the OBS API as before, or do a
shallow clone and inspect the last commit of a package and determine if
it was updated since the last execution.

This commit further moves logic away from cli.py. An integration test
suite would be required to have meaningful tests. That's planned for a
later time.